Merge/next

+++ cmd/lotus-seed/genesis.go
@@ -130,8 +130,9 @@ var genesisAddMinerCmd = &cli.Command{
 			//Balance: initBalance,
 			template.Accounts = append(template.Accounts, genesis.Actor{
 				Type:    genesis.TAccount,
-				Balance: big.Mul(big.NewInt(50_000_000), big.NewInt(build.FilecoinPrecision)),
-				Meta:    (&genesis.AccountMeta{Owner: miner.Owner}).ActorMeta(),
+				Balance: big.Mul(big.NewInt(50_000_000), big.NewInt(int64(build.FilecoinPrecision))),
+				//Balance: initBalance,
+				Meta: (&genesis.AccountMeta{Owner: miner.Owner}).ActorMeta(),
 			})
 		}
